refactor: rename backend folder
