+++ src/backend/movements/impl/bishopMovement.js
