+++ src/backend/movements/straightLineMovement.js
