+++ src/backend/pieces/blackPieceFactory.js
