+++ src/backend/pieces/pieceFactory.js
