+++ src/backend/pieces/whitePieceFactory.js
