+++ src/backend/registry.js
